Migrate to UIWindowScene lifecycle
Add UIApplicationSceneManifest to Info.plist and introduce SceneDelegate
to adopt the UIWindowScene lifecycle and silence the 'UIScene lifecycle
will soon be required' runtime warning.

Changes:
- SceneDelegate.swift: new UIWindowSceneDelegate implementation
  - sceneDidBecomeActive: start ScreenSaverManager (replaces applicationDidBecomeActive)
  - sceneWillResignActive: post .disableScreenSaver notification
  - scene(_:openURLContexts:) + scene(_:willConnectTo:options:): handle
    openhab:// deep links and certificate file imports (replaces
    application(_:open:options:))
- Info.plist: add UIApplicationSceneManifest pointing to SceneDelegate
  and the Main storyboard; UIApplicationSupportsMultipleScenes = false
- AppDelegate.swift:
  - remove var window (owned by SceneDelegate now)
  - remove applicationWillResignActive/DidEnterBackground/WillEnterForeground/
    DidBecomeActive/WillTerminate lifecycle extension
  - remove application(_:open:options:) URL handler
  - remove ScreenSaverManager setup from performDeferredSetup()
  - add application(_:configurationForConnecting:options:) and
    application(_:didDiscardSceneSessions:) for scene management
  - promote notificationDelegate from private to internal so
    SceneDelegate can reach it
- NotificationCenterDelegateImpl.swift: replace
  AppDelegate.appDelegate.window with UIApplication.shared.firstKeyWindow
  (window is no longer on AppDelegate)
- SelectionView.swift, SitemapNavigationView.swift: remove stray
  print() calls in #Preview blocks
